requirements.txt
# OpenCV-only version (no MediaPipe dependency)
opencv-python>=4.5.0
tensorflow>=2.8.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
Pillow>=8.0.0